+++ src/main/java/com/playtomic/tests/wallet/api/WalletController.java
@@ -11,7 +11,6 @@ import com.playtomic.tests.wallet.usecases.dtos.TopupBalanceUseCaseResponse;
 import lombok.AllArgsConstructor;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
-import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 
@@ -20,6 +19,7 @@ import org.springframework.web.bind.annotation.*;
 @AllArgsConstructor
 public class WalletController {
   private static final Logger log = LoggerFactory.getLogger(WalletController.class);
+  public static final String USER_ID = "foo";
 
   private final GetWalletUseCase getWalletUseCase;
   private final TopupBalanceUseCase topupBalanceUseCase;
@@ -28,7 +28,7 @@ public class WalletController {
   public WalletDto getWallet() {
     // TODO add auth interceptor to capture user ID from JWT, and check permissions
     log.info("GET wallet");
-    var input = new GetWalletUseCaseRequest("foo");
+    var input = new GetWalletUseCaseRequest(USER_ID);
     var output = getWalletUseCase.process(input);
     return new WalletDto(output.balance(), output.currency());
   }
@@ -38,14 +38,14 @@ public class WalletController {
     // TODO add auth interceptor to capture user ID from JWT, and check permissions
     log.info("POST wallet/topup: %s".formatted(request));
     var output = topupBalanceUseCase.process(mapRequest(request));
-    if (output.succesful()) {
+    if (output.successful()) {
       return ResponseEntity.ok(mapResponse(output));
     }
     return ResponseEntity.badRequest().body(mapResponse(output));
   }
 
   private TopupBalanceUseCaseRequest mapRequest(TopupRequest request) {
-    return new TopupBalanceUseCaseRequest(request.amount(), request.currency());
+    return new TopupBalanceUseCaseRequest(USER_ID, request.amount(), request.currency(), request.creditCardNumber(), request.idempotencyId());
   }
 
   private TopupResponse mapResponse(TopupBalanceUseCaseResponse output) {