+++ src/main/java/com/playtomic/tests/wallet/usecases/dtos/TopupBalanceUseCaseResponse.java
@@ -1,3 +1,13 @@
 package com.playtomic.tests.wallet.usecases.dtos;
 
-public record TopupBalanceUseCaseResponse(boolean succesful) {}
+public record TopupBalanceUseCaseResponse(boolean successful, String errorMessage, String id) {
+
+    public static TopupBalanceUseCaseResponse unsuccessful(String msg) {
+        return new TopupBalanceUseCaseResponse(false, msg, null);
+    }
+
+    public static TopupBalanceUseCaseResponse successful(String paymentId) {
+        return new TopupBalanceUseCaseResponse(true, null, paymentId);
+    }
+
+}